[GTK] timeline is not updated after few seconds when mouse hovers on controls https://bugs.webkit.org/show_bug.cgi?id=149111
Reviewed by Philippe Normand.

Source/WebCore:

Timeline is not updated if controlsAreHidden is true. The problem here is that
the function does not mean actually 'hidden' since it only checkes 'show' and 'hidden'
class existences. The panel's visibility are not only controlled by the two classes,
but also by video::-webkit-media-controls-panel:hover. The panel could be visible
by setting the pseudo hover class. So we need to check if panel is hovered as well in controlsAreHidden().

Test: media/media-controls-timeline-updates-when-hovered.html

* Modules/mediacontrols/mediaControlsBase.js:
(Controller.prototype.controlsAreHidden):

LayoutTests:

* media/media-controls-timeline-updates-when-hovered-expected.txt: Added.
* media/media-controls-timeline-updates-when-hovered.html: Added.


Identifier: 167487@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190053 268f45cc-cd09-0410-ab3c-d52691b4dbfc