Scrolling a overflow: scroll region makes find overlay holes stick to the edge of the region https://bugs.webkit.org/show_bug.cgi?id=149572 <rdar://problem/13093602>
Reviewed by Simon Fraser.

Test: fast/text/mark-matches-overflow-clip.html

* rendering/InlineTextBox.cpp:
(WebCore::InlineTextBox::paintTextMatchMarker):
Clip the rendered marker rect to the overflow rect.

* testing/Internals.cpp:
(WebCore::Internals::dumpMarkerRectsForNode):
* testing/Internals.h:
* testing/Internals.idl:
Add a internals function to dump the rendered rects for a marker.

* fast/text/mark-matches-overflow-clip-expected.txt: Added.
* fast/text/mark-matches-overflow-clip.html: Added.
Add a test ensuring that a find match that is clipped out by overflow
ends up with a clipped rendered rect.


Identifier: 167673@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@190254 268f45cc-cd09-0410-ab3c-d52691b4dbfc